+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/RegisterServiceSubscribersPassTest.php
@@ -402,8 +402,8 @@ class RegisterServiceSubscribersPassTest extends TestCase
         (new AutowirePass())->process($container);
 
         $expected = [
-            'some.service' => new ServiceClosureArgument(new TypedReference('some.service', 'stdClass')),
-            'some_service' => new ServiceClosureArgument(new TypedReference('stdClass $some_service', 'stdClass')),
+            'some.service' => new ServiceClosureArgument(new TypedReference('stdClass $someService', 'stdClass')),
+            'some_service' => new ServiceClosureArgument(new TypedReference('stdClass $someService', 'stdClass')),
             'another_service' => new ServiceClosureArgument(new TypedReference('stdClass $anotherService', 'stdClass')),
         ];
         $this->assertEquals($expected, $container->getDefinition((string) $locator->getFactory()[0])->getArgument(0));