+++ src/Symfony/Component/RateLimiter/Policy/Window.php
@@ -71,15 +71,13 @@ final class Window implements LimiterStateInterface
         return $this->maxSize - $this->hitCount;
     }
 
-    public function calculateTimeForTokens(int $tokens): int
+    public function calculateTimeForTokens(int $tokens, float $now): int
     {
         if (($this->maxSize - $this->hitCount) >= $tokens) {
             return 0;
         }
 
-        $cyclesRequired = ceil($tokens / $this->maxSize);
-
-        return $cyclesRequired * $this->intervalInSeconds;
+        return (int) ceil($this->timer + $this->intervalInSeconds - $now);
     }
 
     public function __serialize(): array