+++ src/Symfony/Bundle/TwigBundle/CHANGELOG.md
@@ -9,6 +9,11 @@ CHANGELOG
    (check `FileExtensionEscapingStrategy::guess()` for inspiration) and reference it using
    the `twig.autoescape_service` option instead
 
+6.4
+---
+
+ * Allow omitting the `autoescape_service_method` option when `autoescape_service` is set to an invokable service id
+
 6.3
 ---