+++ src/Symfony/Component/Security/Core/CHANGELOG.md
@@ -1,11 +1,18 @@
 CHANGELOG
 =========
 
+
 7.0
 ---
 
  * Remove the `Security` class, use `Symfony\Bundle\SecurityBundle\Security` instead
  * Require explicit argument when calling `TokenStorage::setToken()`
+
+6.4
+---
+
+ * Make `PersistentToken` immutable
+ * Deprecate accepting only `DateTime` for `TokenProviderInterface::updateToken()`, use `DateTimeInterface` instead
 
 6.3
 ---